fix: 修复圆角在缩放时候的大小

+++ src/project_graph/paint/paint_utils.py
@@ -182,7 +182,7 @@ class PainterUtils:
         fill_color: QColor,
         stroke_color: QColor,
         stroke_width: int,
-        radius: int = 0,
+        radius: float = 0,
     ):
         """
         绘制一个矩形，左上角坐标为left_top，宽为width，高为height，填充色为fill_color，边框色为stroke_color
@@ -201,7 +201,12 @@ class PainterUtils:
         painter.setBrush(fill_color)
         painter.setRenderHint(QPainter.Antialiasing)
         painter.drawRoundedRect(
-            int(left_top.x), int(left_top.y), int(width), int(height), radius, radius
+            int(left_top.x),
+            int(left_top.y),
+            int(width),
+            int(height),
+            int(radius),
+            int(radius),
         )
         # HACK: 这个数字转int有隐患，OverflowError: argument 3 overflowed: value must be in the range -2147483648 to 2147483647
         painter.setPen(QColor(0, 0, 0, 0))
